src/test/java/chaomai/ckdtree/snapshot2/CKDTreeMapTest.java
package chaomai.ckdtree.snapshot2;

import org.junit.Assert;
import org.junit.Test;

/**
 * Created by chaomai on 11/22/15.
 */
public class CKDTreeMapTest {
  int dimensionSteps = 1;
  int sampleSteps = 20000;
  int threadsSteps = 10;
  int rounds = 5;
  double delta = 0.001;
  boolean isVerbose = true;

  @Test
  public void testConstructor() throws Exception {
    CKDTreeMap<Integer> ckd = new CKDTreeMap<>(1);
    InternalNode<Integer> root = ckd.root;

    Assert.assertEquals(Double.POSITIVE_INFINITY, root.key[0], delta);
    Assert.assertEquals(Double.POSITIVE_INFINITY, root.left.key[0], delta);
    Assert.assertEquals(Double.POSITIVE_INFINITY, root.right.key[0], delta);
  }

  private void searchDummy() {
    CKDTreeMap<Integer> ckd = new CKDTreeMap<>(1);
    double[] k = {Double.POSITIVE_INFINITY};

    SearchRes<Integer> res1 = ckd.search(k);

    Assert.assertTrue(ckd.contains(k));
  }

  @Test
  public void testSearch() throws Exception {
    searchDummy();
  }
}